Cleaning up old component, removing new navbar
